requirements.txt
# Core ML (Lightweight - Python 3.7 compatible)
scikit-learn==1.0.2
pandas==1.3.5
numpy==1.21.6
joblib==1.1.0

# Experiment Tracking
mlflow==1.30.0

# Hyperparameter Tuning
optuna==3.0.0

# API Serving
fastapi==0.95.0
uvicorn==0.20.0
pydantic==1.10.7

# Data Versioning
dvc==2.58.0

# Utilities
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.0

# Visualization
matplotlib==3.5.3
seaborn==0.12.2

# Testing
pytest==7.4.3
httpx==0.26.0
